api/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
ultralytics
aioboto3==12.1.0
aiobotocore==2.8.0
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
boto3==1.33.1
botocore==1.33.1
certifi==2023.11.17
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
fonttools==4.47.0
frozenlist==1.4.1
graphviz==0.20.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
imageio==2.33.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.2
multidict
networkx==3.2.1
orjson==3.9.10
packaging==23.2
pillow==10.2.0
plotly==5.18.0
pydantic==2.5.3
pydantic-extra-types==2.4.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.8.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tenacity==8.2.3
tifffile==2023.12.9
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4